+++ lib/t_container.mli
@@ -46,6 +46,33 @@ module type Extensions1 = sig
   val any : predicates:('a -> bool) t -> 'a -> bool
   (** [any ~predicates x] tests [x] against [predicates] until one
       returns [true], or all return [false]. *)
+
+  val one : 'a t -> 'a Or_error.t
+  (** [one xs] returns [Ok x] if [xs] is a list containing only [x],
+      and an error otherwise.
+
+      Examples:
+
+      {[
+        T_list.one []     (* error -- not enough *)
+               one [1]    (* ok 1 *)
+               one [1; 2] (* error -- too many *)
+      ]}
+  *)
+
+  val two : 'a t -> ('a * 'a) Or_error.t
+  (** [two xs] returns [Ok (x, y)] if [xs] is a list containing only [x]
+      and [y] in that order, and an error otherwise.
+
+      Examples:
+
+      {[
+        T_list.two []        (* error -- not enough *)
+               two [1]       (* error -- not enough *)
+               two [1; 2]    (* ok (1, 2) *)
+               two [1; 2; 3] (* error -- too many *)
+      ]}
+  *)
 end
 
 module Extend1 (C : Container.S1) : Extensions1 with type 'a t := 'a C.t