+++ lib/t_list.mli
@@ -51,30 +51,3 @@ val prefixes : 'a t -> 'a t t
       prefixes [1; 2; 3] (* [ [ 1 ]; [ 1; 2 ]; [ 1; 2; 3 ] *)
     ]}
 *)
-
-val one : 'a list -> 'a Or_error.t
-(** [one xs] returns [Ok x] if [xs] is a list containing only [x],
-    and an error otherwise.
-
-    Examples:
-
-    {[
-      one []     (* error -- not enough *)
-      one [1]    (* ok 1 *)
-      one [1; 2] (* error -- too many *)
-    ]}
-*)
-
-val two : 'a list -> ('a * 'a) Or_error.t
-(** [two xs] returns [Ok (x, y)] if [xs] is a list containing only [x]
-    and [y] in that order, and an error otherwise.
-
-    Examples:
-
-    {[
-      two []        (* error -- not enough *)
-      two [1]       (* error -- not enough *)
-      two [1; 2]    (* ok (1, 2) *)
-      two [1; 2; 3] (* error -- too many *)
-    ]}
-*)