+++ CHANGES.md
@@ -1,15 +1,24 @@
 # Scheduled for next release
 
+## Breaking changes
+
 - _Potentially breaking change:_ `Traversable.S0_container` now
   contains `module Elt : Equal.S`, and constrains `type elt` to be
   equal to `Elt.t`.  This reflects the situation in
   `Basic_container0`, and shouldn't break any code using
   `Make_container0`, but may cause custom-built modules to fail to
   type-check.
-- Added `Traversable.Chain0`, a functor for combining two
+
+## New features
+
+- Add `Traversable.Chain0`, a functor for combining two
   `S0_container` instances together for nested traversal.
+- Add `T_fn.disj` to go with `T_fn.conj`.
+- Add `Filter_mappable`, which generalises `List.filter_map`.
+
+## Other
+
 - Improve API documentation.
-- Added `T_fn.disj` to go with `T_fn.conj`.
 
 # v0.1.3 (2018-12-13)